+++ orm/model_scan.go
@@ -43,6 +43,22 @@ func (valuesModel) AfterInsert(_ DB) error {
 	return nil
 }
 
+func (valuesModel) BeforeUpdate(_ DB) error {
+	return nil
+}
+
+func (valuesModel) AfterUpdate(_ DB) error {
+	return nil
+}
+
+func (valuesModel) BeforeDelete(_ DB) error {
+	return nil
+}
+
+func (valuesModel) AfterDelete(_ DB) error {
+	return nil
+}
+
 func (m valuesModel) ScanColumn(colIdx int, _ string, b []byte) error {
 	return types.Scan(m.values[colIdx], b)
 }